requirements.txt
asgiref==3.7.2
Django==4.2.6
sqlparse==0.4.4
django-cors-headers==4.2.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-filter==23.3
drf-yasg==1.21.7
executing==1.2.0
psycopg2==2.9.9
autopep8==2.0.4
autoflake==2.2.1
flake8==6.1.0
Pillow